requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
ortools>=9.8.0
google-generativeai>=0.3.0
